Added test cases
