+++ packages/participants-svc/test/unit/index.test.ts
@@ -1,150 +1,30 @@
-// Service.test.ts
+// index.test.ts
+import { Service } from '../../src/application/service';
 
-import { ConsoleLogger, ILogger } from '@mojaloop/logging-bc-public-types-lib';
-import {
-  AccountsBalancesAdapterMock,
-  AuditClientMock,
-  AuthorizationClientMock,
-  MemoryConfigClientMock,
-  MemoryMessageConsumer,
-  MemoryMessageProducer,
-  ParticipantsRepoMock
-} from "@mojaloop/participants-bc-shared-mocks-lib";
-import { MetricsMock } from '@mojaloop/platform-shared-lib-observability-types-lib';
-import { LoginHelper } from '@mojaloop/security-bc-client-lib';
-import { ParticipantsEventHandler } from '../../src/application/event_handler';
-import { ParticipantAggregate } from '../../src/domain/participant_agg';
-import { UnauthorizedError } from '@mojaloop/security-bc-public-types-lib';
+jest.mock('../../src/application/service', () => ({
+  Service: {
+    start: jest.fn(),
+    stop: jest.fn(),
+  },
+}));
 
+describe('Index', () => {
 
-const hasPrivilege = true;
-const AUTH_N_SVC_BASEURL = process.env["AUTH_N_SVC_BASEURL"] || "http://localhost:3201";
-const AUTH_N_SVC_TOKEN_URL = AUTH_N_SVC_BASEURL + "/token";
-const SVC_CLIENT_ID = process.env["SVC_CLIENT_ID"] || "participants-bc-participants-svc";
-const SVC_CLIENT_SECRET = process.env["SVC_CLIENT_SECRET"] || "superServiceSecret";
-const AUTH_TOKEN = "bearer: MOCKTOKEN";
-const authTokenUrl = "mocked_auth_url";
+    beforeAll(()=> {
+        (Service.start as jest.Mock).mockImplementation(() => Promise.resolve());
+    })
 
-const logger: ILogger = new ConsoleLogger();
-const repoPartMock = new ParticipantsRepoMock();
-const accAndBalAdapterMock = new AccountsBalancesAdapterMock();
-const auditClientMock = new AuditClientMock(logger);
-const authZClientMock = new AuthorizationClientMock(logger, hasPrivilege);
-const msgProducerMock = new MemoryMessageProducer(logger);
-const msgConsumerMock = new MemoryMessageConsumer();
-const metricsMock = new MetricsMock();
-const configClientMock = new MemoryConfigClientMock(logger, authTokenUrl);
-
-
-describe('Participant Service Test', () => {
-
-  const loginHelper = new LoginHelper(AUTH_N_SVC_TOKEN_URL, logger);
-  loginHelper.setAppCredentials(SVC_CLIENT_ID, SVC_CLIENT_SECRET);
-
-  const participantAggMock: ParticipantAggregate = new ParticipantAggregate(
-    configClientMock,
-    repoPartMock,
-    accAndBalAdapterMock,
-    auditClientMock,
-    authZClientMock,
-    msgProducerMock,
-    metricsMock,
-    logger
-  );
-
-  const mockedEventHandler: ParticipantsEventHandler = new ParticipantsEventHandler(
-    msgConsumerMock,
-    participantAggMock,
-    loginHelper,
-    logger
-  )
-
-  beforeAll(async () => {
-
-  });
-
-  afterAll(async () => {
+  beforeEach(() => {
+    // Clear the mock calls before each test
     jest.clearAllMocks();
   });
 
-  test("should be able to run start and init all variables", async () => {
-    // Arrange
-    const spyConsumerSetTopics = jest.spyOn(msgConsumerMock, "setTopics");
-    const spyConsumerConnect = jest.spyOn(msgConsumerMock, "connect");
-    const spyConsumerStart = jest.spyOn(msgConsumerMock, "connect");
-
-
+  test('should start the service', async () => {
     // Act
-    await mockedEventHandler.start();
+    await import('../../src/application/index');
 
     // Assert
-    expect(spyConsumerSetTopics).toBeCalledTimes(1);
-    expect(spyConsumerConnect).toBeCalledTimes(1);
-    expect(spyConsumerStart).toBeCalledTimes(1);
-
-  });
-
-  test("should be able to run stop", async () => {
-    // Arrange
-    const spyConsumerStopMethod = jest.spyOn(msgConsumerMock, "stop");
-
-    // Act
-    await mockedEventHandler.stop();
-
-    // Assert
-    expect(spyConsumerStopMethod).toBeCalledTimes(1);
-
-  });
-
-  test("should throw an UnauthorizedError if getToken returns null", async () => {
-
-    // Arrange
-    const fakeToken = null;
-    const getTokenSpy = jest.spyOn(loginHelper, "getToken").mockResolvedValueOnce(fakeToken as any);
-
-    // Act & Assert
-    await expect(mockedEventHandler["_getServiceSecContext"]()).rejects.toThrow(UnauthorizedError);
-    expect(getTokenSpy).toBeCalledTimes(1);
-  });
-
-  test("should return a CallSecurityContext object with the correct properties", async () => {
-    // Arrange
-    const fakeToken = {
-      payload: {
-        azp: "mocked_client_id",
-        platformRoles: ["mocked_role_id"],
-      },
-      accessToken: "mocked_access_token",
-    };
-    const getTokenSpy = jest.spyOn(loginHelper, "getToken").mockResolvedValueOnce(fakeToken as any);
-
-    // Act
-    const secCtx = await mockedEventHandler["_getServiceSecContext"]();
-
-    // Assert
-    expect(secCtx.clientId).toBe(fakeToken.payload.azp);
-    expect(secCtx.accessToken).toBe(fakeToken.accessToken);
-    expect(secCtx.platformRoleIds).toBe(fakeToken.payload.platformRoles);
-    expect(secCtx.username).toBeNull();
-    expect(getTokenSpy).toBeCalledTimes(1);
-  });
-
-
-  test("should call handleSettlementMatrixSettledEvt() method of ParticipantAggregate when message.msgName is SettlementMatrixSettledEvt", async () => {
-    // Arrange
-    const fakeMessage = {
-      msgName: "SettlementMatrixSettledEvt",
-      msgKey: "mocked_key",
-      msgId: "mocked_id",
-    };
-    const handleSettlementMatrixSettledEvtSpy = jest.spyOn(participantAggMock, "handleSettlementMatrixSettledEvt");
-
-    // Act
-    await mockedEventHandler["_msgHandler"](fakeMessage as any);
-    
-    // Assert
-    expect(handleSettlementMatrixSettledEvtSpy).toBeCalledTimes(1);
-    
+    expect(Service.start).toHaveBeenCalled();
   });
 
 });
